Resolve "rework authentification waiter"

Closes #27

See merge request uib/opsi-configed!19

+++ src/main/java/de/uib/utilities/thread/WaitingWorker.java
@@ -50,7 +50,7 @@ public class WaitingWorker extends SwingWorker<Void, Long> {
 
 		// startAnotherProcess()
 
-		long timeStepMillis = (long) 500;
+		long timeStepMillis = 500;
 
 		logging.debug(this, " doInBackground waitingMillis " + waitingSleeper.getWaitingMillis());
